Struct rs3cache_core::buf::Buffer

pub struct Buffer<T> where
    T: AsRef<[u8]>,  { /* fields omitted */ }

Contains a buffer, exposing various reads on it.

Panics

In general, these methods will panic if they attempt to read beyond end of file.

Implementations

impl<T> Buffer<T> where
T: AsRef<[u8]>,

pub fn new(file: T) -> Self

Constructor for Buffer.

pub fn read_array<const LENGTH: usize>(&mut self) -> [u8; LENGTH]

pub fn read_byte(&mut self) -> i8

Reads a byte as signed 8-bit integer. Wraps Buf::get_i8()

pub fn read_unsigned_byte(&mut self) -> u8

Reads a byte as unsigned 8-bit integer. Wraps Buf::get_u8()

pub fn read_int(&mut self) -> i32

Reads four bytes as an 32-bit signed integer. Wraps Buf::get_i32()

pub fn read_unsigned_int(&mut self) -> u32

Reads four bytes as an 32-bit unsigned integer. Wraps Buf::get_u32()

pub fn read_unsigned_short(&mut self) -> u16

Reads two bytes as an 16-bit unsigned integer. Wraps Buf::get_u16()

pub fn read_short(&mut self) -> i16

Reads two bytes as an 16-bit signed integer. Wraps Buf::get_i16()

pub fn read_smart32(&mut self) -> Option<u32>

Reads two or four unsigned bytes as an 32-bit unsigned integer.

pub fn read_unsigned_smart(&mut self) -> u16

Reads one or two unsigned bytes as an 16-bit unsigned integer.

pub fn read_decr_smart(&mut self) -> Option<u16>

Reads either one or two bytes.

pub fn read_masked_data(&mut self) -> Vec<(Option<u32>, Option<u32>)>

Reads masked data.

pub fn read_smarts(&mut self) -> u32

Reads a multiple of two bytes as an 32-bit unsigned integer.

pub fn read_bitflags(&mut self) -> [bool; 8]

Reads one byte, returning 8 boolean bitflags.

pub fn read_string(&mut self) -> String

Reads a 0-terminated String from the buffer, similar to an OsString. The bytes must be valid 1-byte UTF-8.

pub fn read_padded_string(&mut self) -> String

Reads a 0-start and 0-terminated String from the buffer, similar to OsString. The bytes must be valid 1-byte UTF-8. The leading 0 is discarded.

pub fn read_n_bytes(&mut self, count: usize) -> Vec<u8>

Reads count bytes from the buffer.

pub fn read_3_unsigned_bytes(&mut self) -> u32

Reads three unsigned bytes as an 32-bit unsigned integer.

pub fn read_rgb(&mut self) -> [u8; 3]

Reads three unsigned bytes , returning a [red, blue, green] array.

pub fn read_masked_index(&mut self) -> u16

Reads two obfuscated bytes.

pub fn remaining(&mut self) -> usize

Wraps Buf::remaining().

pub fn seek(&mut self, pos: SeekFrom) -> Result<u64, Error>

Go to some offset in the buffer. Wraps Cursor::seek().

pub fn remainder(&mut self) -> Vec<u8>

Returns the remainder of the buffer.