Trait rs3cache_core::buf::BufExtra

pub trait BufExtra: Buf {
    fn get_array<const LENGTH: usize>(&mut self) -> [u8; LENGTH] { ... }

    fn get_smart32(&mut self) -> Option<u32> { ... }

    fn get_unsigned_smart(&mut self) -> u16 { ... }

    fn get_decr_smart(&mut self) -> Option<u16> { ... }

    fn get_masked_data(&mut self) -> Vec<(Option<u32>, Option<u32>)> { ... }

    fn get_smarts(&mut self) -> u32 { ... }

    fn get_bitflags(&mut self) -> [bool; 8] { ... }

    fn get_string(&mut self) -> String { ... }

    fn get_padded_string(&mut self) -> String { ... }

    fn get_rgb(&mut self) -> [u8; 3] { ... }

    fn get_masked_index(&mut self) -> u16 { ... }
}

Provided methods

fn get_array<const LENGTH: usize>(&mut self) -> [u8; LENGTH]

fn get_smart32(&mut self) -> Option<u32>

Reads two or four unsigned bytes as an 32-bit unsigned integer.

fn get_unsigned_smart(&mut self) -> u16

Reads one or two unsigned bytes as an 16-bit unsigned integer.

fn get_decr_smart(&mut self) -> Option<u16>

Reads either one or two bytes.

fn get_masked_data(&mut self) -> Vec<(Option<u32>, Option<u32>)>

Reads masked data.

fn get_smarts(&mut self) -> u32

Reads a multiple of two bytes as an 32-bit unsigned integer.

fn get_bitflags(&mut self) -> [bool; 8]

Reads one byte, returning 8 boolean bitflags.

fn get_string(&mut self) -> String

Reads a 0-terminated String from the buffer

fn get_padded_string(&mut self) -> String

Reads a 0-start and 0-terminated String from the buffer.

fn get_rgb(&mut self) -> [u8; 3]

Reads three unsigned bytes , returning a [red, blue, green] array.

fn get_masked_index(&mut self) -> u16

Reads two obfuscated bytes.